requirements.txt
# FastAPI and server
fastapi==0.123.0
uvicorn[standard]==0.38.0
pydantic==2.12.5
pydantic-settings==2.12.0

# Database
sqlalchemy==2.0.44
asyncpg==0.31.0
psycopg2-binary==2.9.11
greenlet==3.1.1

# Utilities
python-dotenv==1.2.1
python-multipart==0.0.20
httpx==0.28.1
python-dateutil==2.9.0

# AI/ML - LangChain and Azure OpenAI
langchain==1.1.2
langchain-openai==1.1.0
langgraph==1.0.4
azure-identity==1.25.1
openai==2.8.1
